+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/AtmosphericScattering/Fog.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Diagnostics;
+using UnityEngine.Serialization;
 
 namespace UnityEngine.Rendering.HighDefinition
 {
@@ -8,7 +9,7 @@ namespace UnityEngine.Rendering.HighDefinition
     /// </summary>
     [Serializable, VolumeComponentMenu("Fog")]
     [HelpURL(Documentation.baseURL + Documentation.version + Documentation.subURL + "Override-Fog" + Documentation.endURL)]
-    public class Fog : VolumeComponent
+    public class Fog : VolumeComponentWithQuality
     {
         /// <summary>Enable fog.</summary>
         [Tooltip("Enables the fog.")]
@@ -76,30 +77,65 @@ namespace UnityEngine.Rendering.HighDefinition
 
         /// <summary>Controls which method to use to control the performance and quality of the volumetric fog.</summary>
         /// <remarks>Balance mode allows you to use a performance-oriented approach to define the quality of the volumetric fog. Manual mode gives you access to the internal set of properties which directly control the effect.</remarks>
+        public FogControl fogControlMode
+        {
+            get
+            {
+                if (!UsesQualitySettings())
+                    return m_FogControlMode.value;
+                else
+                    return GetLightingQualitySettings().Fog_ControlMode[(int)quality.value];
+            }
+            set { m_FogControlMode.value = value; }
+        }
+        [SerializeField, FormerlySerializedAs("fogControlMode")]
         [Tooltip("Controls which method to use to control the performance and quality of the volumetric fog. Balance mode allows you to use a performance-oriented approach to define the quality of the volumetric fog. Manual mode gives you access to the internal set of properties which directly control the effect.")]
-        public FogControlParameter fogControlMode = new FogControlParameter(FogControl.Balance);
+        private FogControlParameter m_FogControlMode = new FogControlParameter(FogControl.Balance);
+
         /// <summary>Stores the resolution of the volumetric buffer (3D texture) along the x-axis and y-axis relative to the resolution of the frame buffer.</summary>
         [Tooltip("Stores the resolution of the volumetric buffer (3D texture) along the x-axis and y-axis relative to the resolution of the frame buffer.")]
         public ClampedFloatParameter screenResolutionPercentage = new ClampedFloatParameter(optimalFogScreenResolutionPercentage, minFogScreenResolutionPercentage, maxFogScreenResolutionPercentage);
         /// <summary>Number of slices of the volumetric buffer (3D texture) along the camera's focal axis.</summary>
         [Tooltip("Number of slices of the volumetric buffer (3D texture) along the camera's focal axis.")]
         public ClampedIntParameter volumeSliceCount = new ClampedIntParameter(64, 1, maxFogSliceCount);
+        
         /// <summary>Defines the performance to quality ratio of the volumetric fog. A value of 0 being the least resource-intensive and a value of 1 being the highest quality.</summary>
         /// <remarks>Try to minimize this value to find a compromise between quality and performance. </remarks>
+        public float volumetricFogBudget
+        {
+            get
+            {
+                if (!UsesQualitySettings())
+                    return m_VolumetricFogBudget.value;
+                else
+                    return GetLightingQualitySettings().Fog_Budget[(int)quality.value];
+            }
+            set { m_VolumetricFogBudget.value = value; }
+        }
+        [SerializeField, FormerlySerializedAs("volumetricFogBudget")]
         [Tooltip("Defines the performance to quality ratio of the volumetric fog. A value of 0 being the least resource-intensive and a value of 1 being the highest quality.")]
-        public ClampedFloatParameter volumetricFogBudget = new ClampedFloatParameter(0.25f, 0.0f, 1.0f);
+        private ClampedFloatParameter m_VolumetricFogBudget = new ClampedFloatParameter(0.25f, 0.0f, 1.0f);
+
         /// <summary>Controls how Unity shares resources between Screen (XY) and Depth (Z) resolutions.</summary>
         /// <remarks>A value of 0 means Unity allocates all of the resources to the XY resolution, which reduces aliasing, but increases noise. A value of 1 means Unity allocates all of the resources to the Z resolution, which reduces noise, but increases aliasing. This property allows for linear interpolation between the two configurations.<remarks>
+        public float resolutionDepthRatio
+        {
+            get
+            {
+                if (!UsesQualitySettings())
+                    return m_ResolutionDepthRatio.value;
+                else
+                    return GetLightingQualitySettings().Fog_DepthRatio[(int)quality.value];
+            }
+            set { m_ResolutionDepthRatio.value = value; }
+        }
+        [SerializeField, FormerlySerializedAs("resolutionDepthRatio")]
         [Tooltip("Controls how Unity shares resources between Screen (XY) and Depth (Z) resolutions.")]
-        public ClampedFloatParameter resolutionDepthRatio = new ClampedFloatParameter(0.5f, 0.0f, 1.0f);
+        public ClampedFloatParameter m_ResolutionDepthRatio = new ClampedFloatParameter(0.5f, 0.0f, 1.0f);
 
         /// <summary>Indicates whether Unity includes or excludes non-directional light types when it evaluates the volumetric fog. Including non-directional lights increases the resource intensity of the effect.</summary>
         [Tooltip("Indicates whether Unity includes or excludes non-directional light types when it evaluates the volumetric fog. Including non-directional lights increases the resource intensity of the effect.")]
         public BoolParameter directionalLightsOnly = new BoolParameter(false);
-
-        /// <summary>Deprecated don't used</summary>
-        [Tooltip("Deprecated don't used")]
-        public BoolParameter filter = new BoolParameter(false); // DO not used, removed in 11.x
 
         internal static bool IsFogEnabled(HDCamera hdCamera)
         {