+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -15,6 +15,7 @@ The version number for this package has increased due to a version update of a r
 - Added a rough distortion frame setting and and info box on distortion materials.
 - Adding support of 4 channel tex coords for ray tracing (case 1265309).
 - Added a help button on the volume component toolbar for documentation.
+- Added quality preset of the new volumetric fog parameters.
 
 ### Fixed
 - Fixed an issue where the Exposure Shader Graph node had clipped text. (case 1265057)